Make all structure tables reference FQCN's instead of primary keys.
This allows us to drastically simplify the logic of the indexer as we don't have to jump through hoops to keep the ID's intact. We can instead just delete an entire file from the index and let the cascading constraints handle the rest. All references to a class will be using its FQCN and become 'dangling' at that point as there is no foreign key. This, however, is no problem as there is no point in trying to guarantee database integrity for a code base index of a code base that can't be guaranteed to be integer itself (think of files being moved around during refactoring).