Let resolvedType return the FQCN if it is not a special type.
This way users that are only interested in the actual FQCN, regardless of what the original type is, can always use the 'resolvedType'. Those that are interested in the original type can use 'type'. 'fqcn' is simply the fully qualified version of the 'type', but will still remain the same as 'type' when the type is a special type such as 'self', 'static', or '$this'.