Don't delete the structure by FQCN when indexing it.
It was already removed by the cascading constraint on the file. This also fixes a classlike suddenly no longer being found if you defined another classlike with the same FQCN in another file (i.e. after copying a file). This happened, most annoyingly, even if you then changed the FQCN in the copied file, until you did a reindex of the original file.