Rewrite caching layer.
Instead of caching the index data adapter, there is now a caching
proxy for the storage and one for the IndexDataAdapter provider
interface. These are not only more reusable, but also further
improve performance as the queries fetching the ID's for a
specific FQCN are now also cached. The problem where they were
not being cleared on a reindex has instead been solved by marking
the appropriate cache ID's as dirty on a reindex.