Fix circular dependency exception when the same FQCN was used twice
If you have the same FQCN twice in your code base, only one of
them will be analyzed when examining classlikes for associations
(parents, traits, interfaces, ...). Having two classes with the
same FQCN is actually an error, but it can happen if you store
files that aren't actually directly part of your code inside
your project folder. A better solution for this is to exclude
those folders from indexing, which will be possible as soon as
project support is implemented. Until that happens, this should
mitigate the circular dependency exception that ensued because
two classlikes with the same name were examined.