Revise type deduction from conditionals when fetching variable types
This distinguishes between types "assured" (e.g. "$a === null") and types
"excluded" (e.g. "$a !== null") by conditionals. Support for comparisons
such as "$a != null" and "$a !== null" has been added. This results in
the following case (preudocode) now being linted correctly:

$a = Foo|null;

if ($a !== null) {
    $a->foo(); // Now recognizes that $a can never be null.
}

Finally, when examining conditionals, the types they "assure"  or
"exclude" due to their conditions are no longer relevant if the variable
is reassigned to something else.