Refactor type deduction off into a separate class
This is very much a work in progress. At the moment the TypeDeducer
is pretty much a god class that has everything it needs dependency
injected. This will be refactored further step by step. There is
now also an explosion of creation methods for classes across
different commands, which will also be refactored in the future.