Reuse pending promises for identical queries
When two identical queries follow closely together, the cache was not
getting hit because the first promise was still pending. This caused two
competing promises to be spawned, awaiting two separate result sets from
the server side, which would both eventually result in the same output.
Instead of spawning another promise, the second query will now wait for
the result of the previous promise.

This should result in less queries being sent to the PHP server,
resulting in less load and improved performance and responsiveness
in general.