Just warn about socket closes
After reading a bit more about sockets, I'm inclined to believe that
sockets closing by themselves is apparently second nature to them.
One end of the socket connection can close the connection at any
point in time, for any reason. Also, there are various other things
that can interfere with the connection beyond the application's control,
such as the operating system, firewalls, ...

For these reasons, and that automatic reconnection on socket closures
now seems to be stable, I've decided to degrade the severity of these
events to a console warning, not bothering the user with them anymore.

Fixes #259