Switch back to indie linter
... but still support disabling linting on save instead of
automatically, by supporting indexing on save instead of continuously.

We need the indie linter instead of the regular linter since indexing
needs to complete before we can get a fresh overview of problems.
Without it, because linter happens to register its onDidStopChanging
event for the buffer before we do, the lint request is sent before
the reindex is sent.