Lint indexed source code instead of buffer source code
Because we buffer source code for editors that are modified in quick
succession, to ensure that large files don't queue a large amount of
indexing processes because the active one can't finish in time, the
LinterProvider might queue a linting event right after indexing of a
file has succeeded, but before the latest version is indexed.

In other words, the following could happen:

  1) User modifies buffer, indexing event is sent.
  2) User modifies buffer whilst indexing is running, second indexing
     event is queued.
  3) First indexing finishes successfully. Event is fired.
  4) LinterProvider picks up event and takes it as signal to start
     linting file.
  5) Linter sends linting request to core.
  6) Second indexing event is fired.

In step 5, the linter was using the current contents of the text buffer
of the editor. Because of this, it was in fact trying to lint the
source code containing changes that hadn't been indexed yet (which
won't happen earlier than step 6). This in turn causes the core to
sometimes complain that a range was out of bounds because no namespace
existed for it (the newly added range had not been indexed yet, so
the core was, in fact, correct).