Add console debug statements to help debug issue
In some rare cases, the indexing progress bar gets stuck and the
ProjectManager no longer wants to index a file because it's
"isBeingIndexed" flag is still set to true. Both of these conditions
occurring simultaneously seems to rule out an issue with either of
them individually. The cause seems to be the promise for reindexing
never resolving or failing for a certain file.

Perhaps this is related to the proxy encountering an invalid response
and not rejecting the response, but in that case there should be a
console warning to indicate it. Hopefully these debug statements will
help us find the cause next time it happens.