Fix structural elements disappearing from indexes sometimes
This can be reproduced by updating a file (e.g. duplicating a line),
saving it (e.g. Ctrl + S) and immediately closing the file
(e.g. Ctrl + W). The timeout of 500 milliseconds then triggered after
the file was closed. This in turn caused Atom to return an empty buffer
for the editor, which was then passed as STDIN to the signatureHelp
command, which then indexed that empty string as new contents, hence
the disappearance of all structural elements.

Why Atom doesn't just throw an error if you try to fetch the buffer of
a closed editor is beyond me, but here we are.

See also https://gitlab.com/php-integrator/core/issues/148