Throw out CachingProxy
It was useful in a time where most requests happened via Atom packages,
but most requests are now internal to core commands and only the
annotations and refactoring package are left. The former only sends
requests on file changes - which resulted in a cache clear already -
and the latter only hits the cache on rare occasions.