Cancel reindex requests instead of internally queueing them
With the queue, the first reindex request for a series of them was
always sent and left to be handled completely, whilst only the last of
any subsequent requests was maintained in order to skip unnecessary
intermediate ones. This way the core did not have to play catch up.

The unfortunate part was that the first reindex request always went
through, even if the core was very busy handling other requests and
could not handle that first request until a while later, after which
that first request might also have been superseded. As the request had
already been sent, however, there was no way to cancel it.

As the core now has request cancellation support, it is simpler to let
all reindex requests pass through to the core and just cancel the ones
that aren't relevant. This automatically allows us to cancel even the
first request, at least if it wasn't being processed yet - but that was
also the case before.

In short, this should improve latency on slower hardware where many
actions are being performed simultaneously.

References #347