Fix changes made before quickly closing file being ignored
We fixed the issue that this resulted in empty contents being sent to
the server, but now any last changes are just ignored because
onDidStopChanging is not triggered anymore after the editor closes. This
change fixes that.

This coincidentally also fixes #336.

Relates to https://gitlab.com/Serenata/Serenata/issues/183

Fixes #336