Fix built-in structural elements disappearing on update sometimes
This happened because the path to the server or core included its
version number. When first indexing a project, stubs for built-in
structural elements are indexed and this path is taken up in the index.
If the server is later updated to a newer version via this package,
the path changes, but the built-ins are never reindexed. This, combined
with the fact that the server prunes removed files every so often,
caused the notorous "Function is not defined or imported anywhere",
"Classlike is not defined or imported anywhere", and so on.

References https://gitlab.com/Serenata/Serenata/issues/185

Fixes #314