Support displaying code lenses returned by server
This replaces the annotations we had with code lenses, which are 
displayed as inline markers that do not interfere with the code itself. 
These look like badges that can be clicked, much like gutter annotations 
that can be clicked.

The advantage of lenses is that they aren't almost invisible in the 
gutter and they support arbitrary commands via the server. The ones 
we're supporting now just open files, but the server could add lenses in 
the future that perform refactoring actions or do other tasks.

This only wires in retrieving them from the server and displaying them 
in the editor. A click handler is already present, but I have yet to 
request that the server execute the command when that happens (and to 
handle the server asking to open a file at a specific location, as a 
result).

References #460