Don't store indexes in project folder by default anymore due to CPU rise
I wanted to place the index inside the .serenata folder, but it turns 
out that is a very bad idea. Atom will start firing massive amounts of 
change requests, due to it watching the database file, every time it is 
modified. We can disable this for our package, but not for other 
packages, which will still receive these events en masse uselessly, 
which not only prevents any other responses from being handled in the 
meantime, it also spikes CPU usage.

Existing indexes will continue to work, this just sets the default for 
new projects to be the php-ide-serenata cache folder. You can still opt 
to save them in the project, by changing it, if you want.

References #460